test: share the output-row rendering and the count sink of the IR tests
